Changed base_address from uint8 to long

+++ omap/kernel/arch/address.h
@@ -6,10 +6,10 @@
 #define address(X, Y) __CONCRETE_ADDRESS(X, Y)
 
 //address of a register
-typedef volatile uint32_t* address;
+typedef volatile long* address;
 
 //base address
-typedef volatile uint8_t base_address;
+typedef volatile long base_address;
 
 //offset to base address
 typedef volatile uint16_t offset;
